% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivosLiquidos.R
\name{ActLiquidos}
\alias{ActLiquidos}
\title{Activos Liquidos.}
\usage{
ActLiquidos(Disp, BCV, BOIFP, BCE, ITVPN)
}
\arguments{
\item{Disp}{Vector que contiene la cuenta Disponibilidades.}

\item{BCV}{Vector que contiene lo dispuesto en Encaje Legal.}

\item{BOIFP}{Vector que contiene activos dispuestos en otros
bancos o instituciones financieras dentro del pais.}

\item{BCE}{Vector que contiene activos dispuestos en bancos
corresponsales en el exterior.}

\item{ITVPN}{Vector que contiene las inversiones en titulos
valores para negociar.}
}
\value{
Si todas los parametros estan definidos entonces la funcion
  devuelve un data frame que contiene los activos liquidos 
  necesarios para la evaluacion de los indicadores contable 
  de liquidez.
}
\description{
\code{ActLiquidos} Devuelve un data frame que continen los 
  Activos Liquidos de Balance.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
